Refactor location and name of some files.
